move changelog; cleanup
