+++ projects/gateway2/krtcollections/pods.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"maps"
 
-	core "github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
-	"google.golang.org/protobuf/proto"
 	"istio.io/api/label"
 	"istio.io/istio/pkg/kube"
 	"istio.io/istio/pkg/kube/kclient"
@@ -13,37 +11,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 )
-
-type GlooResource interface {
-	proto.Message
-	interface {
-		GetMetadata() *core.Metadata
-	}
-}
-
-type ResourceWrapper[T GlooResource] struct {
-	Inner T
-}
-
-func (us ResourceWrapper[T]) ResourceName() string {
-	return krt.Named{
-		Name:      us.Inner.GetMetadata().GetName(),
-		Namespace: us.Inner.GetMetadata().GetNamespace(),
-	}.ResourceName()
-}
-
-func (us ResourceWrapper[T]) Equals(in ResourceWrapper[T]) bool {
-	return proto.Equal(us.Inner, in.Inner)
-}
-
-func (us ResourceWrapper[T]) GetMetadata() *core.Metadata {
-	return us.Inner.GetMetadata()
-}
-
-// equivalent of var _ Interface = Struct{} for generics
-func _genericTypeAssert[T GlooResource]() (krt.ResourceNamer, krt.Equaler[ResourceWrapper[T]]) {
-	return ResourceWrapper[T]{}, ResourceWrapper[T]{}
-}
 
 type NodeMetadata struct {
 	name   string
